Move connection form items into tabs
