+++ src/views/webview-app/components/connect-form/connection-form.less
@@ -46,13 +46,31 @@
 .connection-form-tab-selected::after {
   content: '';
   position: absolute;
+  bottom: 0px;
   height: 4px;
-  bottom: 0;
   left: -1px;
   right: -1px;
+  pointer-events: none;
+  outline: none;
   background-color: #13AA52;
   border-top-left-radius: 3px;
   border-top-right-radius: 3px;
+  animation-name: selectedConnectionTabAnimation;
+  animation-duration: 200ms;
+}
+
+@keyframes selectedConnectionTabAnimation {
+  0% {
+    height: 0px;
+    left: 50%;
+    right: 50%;
+  }
+
+  100% {
+    height: 4px;
+    left: -1px;
+    right: -1px;
+  }
 }
 
 .connection-form-fields {
