+++ src/views/webview-app/components/connect-form/general-tab/authentication/authentication.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
 
-import { ActionTypes, AuthStrategyChangedAction } from '../../../store/actions';
+import { ActionTypes, AuthStrategyChangedAction } from '../../../../store/actions';
 import AUTH_STRATEGIES, {
   AuthStrategies
-} from '../../../connection-model/constants/auth-strategies';
-import FormGroup from '../../form/form-group';
-import FormItemSelect from '../../form/form-item-select';
+} from '../../../../connection-model/constants/auth-strategies';
+import FormGroup from '../../../form/form-group';
+import FormItemSelect from '../../../form/form-item-select';
 import Kerberos from './kerberos';
 import LDAP from './ldap';
 import MongoDBAuth from './mongodb-authentication';
@@ -33,8 +33,6 @@ type props = {
 } & DispatchProps;
 
 class Authentication extends React.Component<props> {
-  static displayName = 'Authentication';
-
   /**
    * Changes an authentication strategy.
    *
