+++ src/views/webview-app/components/connect-form/ssh-tab/ssh-tunnel-password-validation.tsx
@@ -3,21 +3,18 @@ import { connect } from 'react-redux';
 
 import {
   ActionTypes,
-  OnChangeSSHTunnelIdentityFileAction,
   SSHTunnelHostnameChangedAction,
-  SSHTunnelPassphraseChangedAction,
+  SSHTunnelPasswordChangedAction,
   SSHTunnelPortChangedAction,
   SSHTunnelUsernameChangedAction
 } from '../../../store/actions';
 import { AppState } from '../../../store/store';
 import FormInput from '../../form/form-input';
-import FileInputButton from '../../form/file-input-button';
 import FormGroup from '../../form/form-group';
 
 type DispatchProps = {
-  onChangeSSHTunnelIdentityFile: () => void;
   onSSHTunnelHostnameChanged: (sshTunnelHostname: string) => void;
-  onSSHTunnelPassphraseChanged: (sshTunnelPassphrase: string) => void;
+  onSSHTunnelPasswordChanged: (sshTunnelPassword: string) => void;
   onSSHTunnelPortChanged: (sshTunnelPort: number) => void;
   onSSHTunnelUsernameChanged: (sshTunnelUsername: string) => void;
 };
@@ -25,17 +22,14 @@ type DispatchProps = {
 type StateProps = {
   isValid: boolean;
   sshTunnelHostname?: string;
-  sshTunnelIdentityFile?: string[];
-  sshTunnelPassphrase?: string;
+  sshTunnelPassword?: string;
   sshTunnelPort: number;
   sshTunnelUsername?: string;
 };
 
 type props = StateProps & DispatchProps;
 
-class SSHTunnelIdentityFileValidation extends React.Component<props> {
-  static displayName = 'SSHTunnelIdentityFileValidation';
-
+class SSHTunnelPasswordValidation extends React.Component<props> {
   /**
    * Handles sshTunnelHostname change.
    *
@@ -55,21 +49,12 @@ class SSHTunnelIdentityFileValidation extends React.Component<props> {
   };
 
   /**
-   * Handles sshTunnelIdentityFile change.
+   * Handles sshTunnelPassword change.
    *
    * @param {Object} evt - evt.
    */
-  onChangeSSHTunnelIdentityFile = (): void => {
-    this.props.onChangeSSHTunnelIdentityFile();
-  };
-
-  /**
-   * Handles sshTunnelPassphrase change.
-   *
-   * @param {Object} evt - evt.
-   */
-  onSSHTunnelPassphraseChanged = (evt): void => {
-    this.props.onSSHTunnelPassphraseChanged(evt.target.value);
+  onSSHTunnelPasswordChanged = (evt): void => {
+    this.props.onSSHTunnelPasswordChanged(evt.target.value);
   };
 
   /**
@@ -85,14 +70,13 @@ class SSHTunnelIdentityFileValidation extends React.Component<props> {
     const {
       isValid,
       sshTunnelHostname,
-      sshTunnelIdentityFile,
-      sshTunnelPassphrase,
+      sshTunnelPassword,
       sshTunnelPort,
       sshTunnelUsername
     } = this.props;
 
     return (
-      <FormGroup id="sshTunnelIdentityFileValidation">
+      <FormGroup id="sshTunnelPassword">
         <FormInput
           label="SSH Hostname"
           name="sshTunnelHostname"
@@ -104,10 +88,10 @@ class SSHTunnelIdentityFileValidation extends React.Component<props> {
         <FormInput
           label="SSH Tunnel Port"
           name="sshTunnelPort"
-          placeholder="22"
           error={!isValid && sshTunnelPort === undefined}
           changeHandler={this.onSSHTunnelPortChanged}
           value={sshTunnelPort}
+          type="number"
         />
         <FormInput
           label="SSH Username"
@@ -116,19 +100,13 @@ class SSHTunnelIdentityFileValidation extends React.Component<props> {
           changeHandler={this.onSSHTunnelUsernameChanged}
           value={sshTunnelUsername || ''}
         />
-        <FileInputButton
-          label="SSH Identity File"
-          id="sshTunnelIdentityFile"
-          error={!isValid && sshTunnelIdentityFile === undefined}
-          onClick={this.onChangeSSHTunnelIdentityFile}
-          values={sshTunnelIdentityFile}
-        />
         <FormInput
-          label="SSH Passphrase"
-          name="sshTunnelPassphrase"
+          label="SSH Password"
+          name="sshTunnelPassword"
           type="password"
-          changeHandler={this.onSSHTunnelPassphraseChanged.bind(this)}
-          value={sshTunnelPassphrase || ''}
+          error={!isValid && sshTunnelPassword === undefined}
+          changeHandler={this.onSSHTunnelPasswordChanged}
+          value={sshTunnelPassword || ''}
         />
       </FormGroup>
     );
@@ -139,28 +117,24 @@ const mapStateToProps = (state: AppState): StateProps => {
   return {
     isValid: state.isValid,
     sshTunnelHostname: state.currentConnection.sshTunnelHostname,
-    sshTunnelIdentityFile: state.currentConnection.sshTunnelIdentityFile,
-    sshTunnelPassphrase: state.currentConnection.sshTunnelPassphrase,
+    sshTunnelPassword: state.currentConnection.sshTunnelPassword,
     sshTunnelPort: state.currentConnection.sshTunnelPort,
     sshTunnelUsername: state.currentConnection.sshTunnelUsername
   };
 };
 
 const mapDispatchToProps: DispatchProps = {
-  onChangeSSHTunnelIdentityFile: (): OnChangeSSHTunnelIdentityFileAction => ({
-    type: ActionTypes.ON_CHANGE_SSH_TUNNEL_IDENTITY_FILE
-  }),
   onSSHTunnelHostnameChanged: (
     sshTunnelHostname: string
   ): SSHTunnelHostnameChangedAction => ({
     type: ActionTypes.SSH_TUNNEL_HOSTNAME_CHANGED,
     sshTunnelHostname
   }),
-  onSSHTunnelPassphraseChanged: (
-    sshTunnelPassphrase: string
-  ): SSHTunnelPassphraseChangedAction => ({
-    type: ActionTypes.SSH_TUNNEL_PASSPHRASE_CHANGED,
-    sshTunnelPassphrase
+  onSSHTunnelPasswordChanged: (
+    sshTunnelPassword: string
+  ): SSHTunnelPasswordChangedAction => ({
+    type: ActionTypes.SSH_TUNNEL_PASSWORD_CHANGED,
+    sshTunnelPassword
   }),
   onSSHTunnelPortChanged: (
     sshTunnelPort: number
@@ -179,4 +153,4 @@ const mapDispatchToProps: DispatchProps = {
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(SSHTunnelIdentityFileValidation);
+)(SSHTunnelPasswordValidation);
