+++ src/views/webview-app/components/connect-form/ssh-tab/ssh-tunnel.tsx
@@ -3,24 +3,23 @@ import { connect } from 'react-redux';
 
 import { ActionTypes, SSHTunnelChangedAction } from '../../../store/actions';
 import SSH_TUNNEL_TYPES, {
-  SSHTunnelOptions,
+  SSHTunnelOptions
 } from '../../../connection-model/constants/ssh-tunnel-types';
 import FormGroup from '../../form/form-group';
 import FormItemSelect from '../../form/form-item-select';
 import SSHTunnelPasswordValidation from './ssh-tunnel-password-validation';
 import SSHTunnelIdentityFileValidation from './ssh-tunnel-identity-file-validation';
+import { AppState } from '../../../store/store';
 
 type DispatchProps = {
   onSSHTunnelChanged: (sshTunnel: SSH_TUNNEL_TYPES) => void;
 };
 
-type props = {
+type StateProps = {
   sshTunnel: string;
-} & DispatchProps;
+};
 
-class SSHTunnel extends React.Component<props> {
-  static displayName = 'SSHTunnel';
-
+class SSHTunnel extends React.Component<StateProps & DispatchProps> {
   /**
    * Handles SSH tunnel change.
    *
@@ -56,7 +55,7 @@ class SSHTunnel extends React.Component<props> {
           label="SSH Tunnel"
           name="sshTunnel"
           options={SSHTunnelOptions.map((sshTunnelOption) => ({
-            [`${sshTunnelOption.id}`]: sshTunnelOption.title,
+            [`${sshTunnelOption.id}`]: sshTunnelOption.title
           }))}
           changeHandler={this.onSSHTunnelChanged}
           value={sshTunnel}
@@ -67,11 +66,15 @@ class SSHTunnel extends React.Component<props> {
   }
 }
 
+const mapStateToProps = (state: AppState): StateProps => ({
+  sshTunnel: state.currentConnection.sshTunnel
+});
+
 const mapDispatchToProps: DispatchProps = {
   onSSHTunnelChanged: (newSSHTunnel): SSHTunnelChangedAction => ({
     type: ActionTypes.SSH_TUNNEL_CHANGED,
-    sshTunnel: newSSHTunnel,
-  }),
+    sshTunnel: newSSHTunnel
+  })
 };
 
-export default connect(null, mapDispatchToProps)(SSHTunnel);
+export default connect(mapStateToProps, mapDispatchToProps)(SSHTunnel);
