+++ src/views/webview-app/components/connect-form/ssl-tab/ssl-server-client-validation.tsx
@@ -30,8 +30,6 @@ type DispatchProps = {
 type props = StateProps & DispatchProps;
 
 class SSLServerClientValidation extends React.Component<props> {
-  static displayName = 'SSLServerClientValidation';
-
   /**
    * Handles sslCA change.
    */
