+++ src/views/webview-app/connection-model/constants/ssh-tunnel-types.ts
@@ -1,5 +1,5 @@
 // Allowed values for the `sshTunnel` field.
-enum SSL_TUNNEL_TYPES {
+enum SSH_TUNNEL_TYPES {
   /**
    * Do not use SSH tunneling.
    */
@@ -15,23 +15,23 @@ enum SSL_TUNNEL_TYPES {
 }
 
 type SSHTunnelOption = {
-  id: SSL_TUNNEL_TYPES;
+  id: SSH_TUNNEL_TYPES;
   title: string;
 };
 
 export const SSHTunnelOptions: SSHTunnelOption[] = [
   {
-    id: SSL_TUNNEL_TYPES.NONE,
+    id: SSH_TUNNEL_TYPES.NONE,
     title: 'None'
   },
   {
-    id: SSL_TUNNEL_TYPES.USER_PASSWORD,
+    id: SSH_TUNNEL_TYPES.USER_PASSWORD,
     title: 'Use Password'
   },
   {
-    id: SSL_TUNNEL_TYPES.IDENTITY_FILE,
+    id: SSH_TUNNEL_TYPES.IDENTITY_FILE,
     title: 'Use Identity File'
   }
 ];
 
-export default SSL_TUNNEL_TYPES;
+export default SSH_TUNNEL_TYPES;
